Fix events map store error
